refactor(site/src/pages/AgentsPage/components/ChatConversation): extract timeline helpers
